requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
psycopg2-binary
pillow
PyYAML
python-dotenv
opencv-python
ultralytics
pandas
numpy
gunicorn
whitenoise
celery
redis
boto3